Adds the metadata key to configuration at install
In order to be used appropriately with the `:console` backend, Timber
needs to have the `:timber_context` and `:event` metadata keys exposed.
Without these keys, it will not be able to print the added context or
event.

This change adds the necessary `:metadata` key in the configuration file
that exposes the metadata to the formatter.