Handles Hackney asynchronous responses better
First, this moves the `wait_on_request/1` functionality from the
`Timber.HTTPClient` behaviour to the `Timber.LoggerBackends.HTTP` event
handler. Hackney sends its response messages back to the calling
process, which is the event handler in this case. The event handler is
responsible for handling those messages appropriately via the
`handle_info/2` function. However, the `handle_info/2` logic was never
written to capture hackney responses, therefore they would have fallen
through to the last function definition and been ignored.

The result of this was that hackney response messages would have been
lost, and worse they wouldn't be sent again. If hackney sent a :done
message for an asynchronous request and it was discarded by
`handle_info/2`, the next `wait_on_request/1` call inside the
`Timber.HTTPClients.Hackney` module would have looped indefinitely since
the reference it was looking for would never be sent.

Also, the existing `wait_on_request/1` code was too greedy. It had a
final wildcard match that would have matched any inbound message. This
means that it would have matched incoming logs, system notices, and
anything else that wasn't one of the hackney responses defined above it.
Since those messages were matched by the wildcard, they were removed
from the process message box and handed into the block's logic.
That match simply called the function in a loop, so those messages were
discarded into the ether.

So, now, instead of calling the `wait_on_request/1` the logic has been
split into two pieces. First, the private function
`handle_hackney_response/2` has been defined in the
`Timber.LoggerBackends.HTTP` module. It takes the message from hackney
as its first parameter and the current state of the event handler as its
second. Its responsibilty is as follows:

  - If the API returns a 401 (unauthorized) response, it raises an error
  causing the event handler to crash. This is expected functionality.
  - Otherwise, it waits until it receives either a done or error message
  to clear the current reference in the state.

The `handle_hackney_response/2` should be called when the message being
matched also matches the current ref.

Actually receiving the message is now the work of `handle_info/2` as
well as the private function `wait_on_request/1` (inside
`Timber.LoggerBackends.HTTP`). `handle_info/2` is part of the sequential
processing of messages by the event handler. When hackney sends a
response message to the event handler and the event handler is not
in a `wait_on_request/1` loop, the message is given to `handle_info/2`
if it matches the current ref. It then calls `handle_hackney_response/2`
to modify the state appropriately.

`wait_on_request/1` is called when the event handler needs to block
until the previous request has finished. It works by looping until the
ref in the state has been cleared. If there is a ref, it iterates over
the current process message box looking for hackney responses for that
ref. If it finds one, it hands it off to `handle_hackney_response/2` to
modify the state appropriately and then loop. If the state is modified
such that the reference is cleared, the next call will exit the loop.
Each call of the receive block will allow up to 5 seconds for a hackney
message to appear before it times out and deems the request's reference
abandoned.

An "abandoned" or "orphaned" request is just that: the request has
exceeded the time the system will wait for it, and as a result, the
system abandons it. Any further messages from hackney about this request
will be passed through to the last function defintiion for `handle_info/2`
and discarded.

As part of this change, it became clear that hackney was integrally tied
ot the event handler. Because the event handler needs to know the
message forms to match on for asynchronous requests. With this in mind,
hackney is now a _required_ dependency of the application. The
configuration field for the HTTP client remains in the HTTP backend's state
but this is only for testing purposes.