Separate context plugs into HTTP and Session (#160)
This introduces the SessionContextPlug and HTTPContextPlug which replace
the monolithic ContextPlug. The HTTPContextPlug captures the HTTPContext
and the SessionContextPlug captures the SessionContext. This is
advantageous for users who do not have a session store set up; those
users encounter a raise during the Plug response phase that happens
outside the package. These users can now remove the SessionContextPlug
but retain the HTTPContextPlug functionality.

The existing ContextPlug module now calls out to both SessionContextPlug
and HTTPContextPlug to maintain compatability. That module will be
removed in version 3.0 of the package and beyond.

Closes #155