Update testing dependencies
The previous inclusion of testing dependencies caused the test
dependencies to erroneously be declared as production dependencies.
This changes the dependencies to only be added for the test environment.

Now, by default, when a value for a dependency constraint is not defined
using an environment variable, the dependency will use the latest
version.

The tests are now run against the gamut of recent Ecto, Phoenix, and
Plug releases. I also removed the "omit" clauses, since they didn't
provide a useful effect.

Fixes #191