Allow Phoenix Controller Blacklisting
This adds the ability to blacklist certain controller/action pairs from
the Phoenix instrumentation system. When the controller/action pair is
added to the application configuration (either from the Mix
configuration file or via the functions provided in the
PhoenixInstrumentation module), instrumentation output is skipped for
that specific pair.

This functionality is implemented using a MapSet to allow for quick
querying of the blacklist. MapSets are backed by maps which should
produce an O(1) lookup time, versus a list which would produce an O(n)
lookup time.

Phoenix template rendering logs also honors this blacklist

Closes #165