Clean up third party integration specifications
This cleans up the way in which versions for third-party integrations
are specified. It's much cleaner and easier to read. It also specifies
the correct version range.

However, I have removed the complex testing matrix for now because the
run time complexity for changes is excessive. We will need to figure out
a way to better support this in the future that gives the same
resilience but lowers the time it takes to get feedback on
compatability.

Closes #191

Closes #218