Fixes instrumentation crash on no controller
The controller render event for Phoenix instrumentation was encountering
an error when the conn being evaluated had no controller or action set.
This occured, for example, when there was no route match and a
`404.html` template was rendered.

I fixed this by checking whether there was a controller and action
present first before taking them and determining their blacklist status.
The logic is slightly more convoluted but prevents the error that is
currently happening.

Closes #230