Fix bad blacklist default in Phoenix integration
The Phoenix integration would default to an empty list instead of an
empty blacklist when trying to fetch the blacklist from the application
environment. This caused an error when the blacklist was later evaluated
using MapSet functions.

I aded a test that ensures the default is set correctly to an empty
MapSet and updated the default appropriately.

Closes #232