Fixed phoenix_render_controller default :start
This changes allows the
`Timber.Integrations.PhoenixInstrumenter.phoenix_render_controller/3`
function to be called with `:start` plus any two additional arguments
and still succeed. This fixes an issue where certain map keys we depend upon
are only available in 1.3 and above. When run on a version of Phoenix
below 1.3, the function would raise with a match failure.

The function will no longer raise with a match faiulre now. Instead it
will simply suppress logging for that event.

Fixes #238