Handle lack of :conn on phoenix_controller_render
We did not appropriately handle calls of
`Timber.Integrations.PhoenixInstrumenter.phoenix_controller_render/3`
where the `:start` event was given a map without a `:conn` key. Now,
when the `:conn` isn't present, we just take the template name and
assume that the event should be logged.

This also clarifies that Phoenix 1.3 is required for full-support of
controller blacklisting.

Fixes #238