Timber.CurrentContext -> Timber.LocalContext
This introduces the module Timber.LocalContext to replace
Timber.CurrentContext. Timber.CurrentContext was repsonsible for
managing the context using the Elixir Logger metadata, which is local to
a process, so the name LocalContext makes more sense. In future commits,
Timber.GlobalContext will be introduced to supplement context stored in
the Logger metadata.