Add global context
This adds support for global context, through the `Timber.GlobalContext` module
as well as through the `Timber.add_context/2` helper function. Global context
will be stored in the OTP application configuration for timber, but if this
proves to be a poor choice, the `Timber.GlobalContext` module allows the storage
locaiton to be modified without changing the API.

The global context is now merged with the local context when log events are
created via `Timber.LogEntry.new/4`. The local context will overwrite the global
context where they conflict.

Closes #196