Fix Keyword list handling bug
This fixes a bug in the Phoenix Channels integration layer. When using Phoenix
1.3+ and the channel payload is a list (not a Keyword list), the
Timber.Integrations.PhoenixInstrumenter.phoenix_channel_receive/3 function would
raise when trying to force the list into a map.

The logic now only coerces a list into a map if the list is a valid Keyword
list (and thus can be represented by a map).

Fixes the bug described in #258