Take unstructured params as-is
Unstructured params in Phoenix Channel payloads were being replaced by an empty
map which wasn't a great experience for users. Users expect the params field to
represent the params that were included in the message. This now accepts params
as-is when they are not a list or map (lists and maps are handled separately).

Closes #258