Prepare 3.0 development
This commit is designed to prepare for 3.0+ development. In Timber
Elixir 3.0+, integration with other libraries will be done through
additional libraries in order to allow for better dependency management
and compilation guarantees. Timber Elixir 3.0+ will be required for
Elixir 1.7+ support.

All of the integration-specific code has been removed except for the
installer code. The installer code needs to be updated to appropriately
update the mix.exs files with the relevant integration libraries and
use the updated module names.

The weird dependency listing logic has been removed from mix.exs.

The README has also been fixed so that the markdown now renders
correctly.

This fixes the .formatter.exs file to look at the root `lib` folder.

This change adds Elixir 1.7.2 to the list of versions that Travis tests
for.

Addresses #298, fixes #209, fixes #162