Change supported Elixir versions to ~> 1.4
The Timber library will now only support Elixir versions 1.4 and higher. To use
Timber with Elixir 1.3 (which support was removed for), please use a release
from the 2.x branch.

Supporting only Elixir 1.4+ will allow us to take advantage of the new Jason
library which is easier to maintain support for.