Provide InMemory Logger backend for testing
This provides the InMemory Logger backend for testing purposes. The InMemory
backend was previously only available as the TestLoggerBackend as a test support
module. This moves the module to the library code (instead of test support code)
so any code dependent on the library can utilize the InMemory logger.

The InMemory logger is only intended to be used for testing dependent libraries,
not for production use.