Use the Jason library for JSON encoding
This changes the JSON encoding paths to use Jason instead of Poison. Like
Poison, Jason is a pure Elixir JSON library, but it has a focus on standards
compliance and speed. It is intended to become the default JSON library on a
number of large Elixir projects including Phoenix. Although Poison 4 was
recently released, Poison lingered for nearly a year without critical bug fixes
being addressed. Poison's interface was also inconsistent and poorly documented
for error cases.

With the switch to Jason, Elixir 1.3 and below can no longer be supported since
Jason relies on Elixir 1.4 and above.

This also removes the ability for the end-developer to inject the JSON library
via configuration. While this is a nice concept in theory, it also introduced
the strong chance the library would not correspond to the expected behavior of
the Timber library logic. Removing this ability ensures we know what is
happening in critical code paths.

As part of this change, I had to address the fact that some tests hard-coded
expected JSON output. Poison output small maps as JSON with a deterministic key
order, so tests passed when using Poison, but the library did not publicize this
as a guarantee. Jason does not have the same determinism for small maps. I
changed these tests to decode the JSON payload and test membership and content.