Fix Credo errors
This fixes files to bring them inline with Credo expectations. It also adds
Credo to the Travis runs.

Additionally, the exception module
`Timber.LoggerBackends.HTTP.TimberAPIKeyInvalid` is renamed to
`Timber.InvalidAPIKeyError` in order to comply with Elixir module naming rules
for exception modules (i.e., they need to end in `Error`). It is also
appropriate that it is its own module rather than being nested inside the HTTP
backend.