Remove `event`, `event.custom`, and `context.custom` namespaces to follow new Timber 2.0 schema changes
Timber 2.0 no longer requires a strict schema, this means that events and contexts can be defined freely without the need to be nested under specific `event`, `event.custom`, or `context.custom` keys.

For example, this `http_response` event went from this:

```json
{
  "level": "info",
  "message": "Order abcd1234 placed",
  "event":
    "custom": {
      "order_placed": {
        "id": "abcd1234",
        "total": 100
      }
    }
  }
}
```

To this:

```json
{
  "level": "info",
  "message": "Order abcd1234 placed",
  "order_placed": {
    "id": "abcd1234",
    "total": 100
  }
}
```

Timber 2.0 is a major update in the Timber platform that does away with a strict or opinionated schema. Instead, it provides a generic framework for logging structured data without any assumptions around events types, names, or structure. That is left at the discretion of the user, giving users free reign as it relates to the events they capture as well as the resulting schema. This has a number of benefits that are reflected in this PR:

1. Custom events are no longer required to be placed in the `event.custom.*` keyspace. This removes the resulting schema awkwardness when using custom events in searches and visualizations.
2. Custom contexts are no longer required to be placed in the `context.custom.*` keyspace for the same reasons mentioned above.
3. Official Timber event structs are no longer required since each user of the Timber library can define events in any structured deems appropriate.
4. Exception event structuring from the exception message itself has been removed in favor of using [`:timber_exceptions`](https://github.com/timberio/timber-elixir-exceptions)

== Backwards compatibility ==

This PR should NOT break backwards compatibility. Methods no longer necessary have been marked as deprecated and will be removed in Timber `4.0.0`