Support new Timber API keys
The new Timber API keys are not source specific, the source ID must be specified separately. This change updates the Timber library to take a `:source_id` configuration paramter in addition to the new API key. If a `:source_id` is supplied it is assumed the API key supplied follows the new format, otherwise the legacy API key method will be used. This ensures we do not break backwards compatibility.