Default formatter to print in JSON
This change removes the fancy `@metadata` formatting for the `Timber.Formatter`, which is a formatter for the `Logger` `:console` backend. The Timber service no longer requires this and it just servers to complicate things. In addition, the recommendation to use the `Timber.Formatter` in development has been removed, the `Timber.Formatter` should only be used in production or staging environments unless the user wants logs printed as JSON.